requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-ai-inference==1.0.0b9
azure-core==1.32.0
azure-identity==1.21.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
distro==1.9.0
fastapi==0.115.11
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.9.0
msal==1.32.0
msal-extensions==1.3.1
openai==1.66.3
passlib==1.7.4
psycopg2==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
